Directory Structure Reorganization
Reorganized files to be cleaner
