+++ src/js/components/Spell/Rotate/AllRotateSelect.jsx
@@ -8,38 +8,32 @@ import Select from '@mui/material/Select';
 
 // Redux
 import { useDispatch, useSelector } from 'react-redux';
+import {
+    selectAllColorShiftRotates
+} from '../../../redux/slices/spells/colorShift';
 import { 
-    selectNearbyRotates,
     selectSpecificRotate,
     setSpecificRotate
-} from '../../redux/slices/currentSelections';
+} from '../../../redux/slices/currentSelections';
 
-export const MatchingRotateSelect = ({ id = 'matching-rotate-select', rootStyle = {} }) => {
+export const AllRotateSelect = ({ id = 'all-rotate-select', rootStyle = {} }) => {
     const dispatch = useDispatch();
 
-    const nearbyRotates = useSelector(selectNearbyRotates);
+    const allRotates = useSelector(selectAllColorShiftRotates);
     const specificRotate = useSelector(selectSpecificRotate);
 
-    const nearbyRotateIds = nearbyRotates.map(rotate => rotate.id);
-    const nearbyRotateCount = nearbyRotates.length;
-
-    const labelText = `(${nearbyRotateCount}) Matching Spell${nearbyRotateCount !== 1 ? 's' : ''}`;
     const labelId = `${id}-label`;
+
     const value = (
         specificRotate
         ? (
-            specificRotate.id === 'fake' ? '' : 
-            (
-                nearbyRotateIds.some(id => id === specificRotate.id)
-                ? specificRotate.id
-                : ''
-            )
+            specificRotate.id === 'fake' ? '' : specificRotate.id
         )
         : ''
     );
 
     const handleRotateChange = (e) => {
-        let selectedSpell = nearbyRotates.find(rotate => rotate.id === e.target.value);
+        let selectedSpell = allRotates.find(rotate => rotate.id === e.target.value);
 
         if (!selectedSpell) {
             selectedSpell = {
@@ -57,17 +51,17 @@ export const MatchingRotateSelect = ({ id = 'matching-rotate-select', rootStyle 
             style={rootStyle}
         >
             <FormControl fullWidth>
-                <InputLabel id={labelId}>{labelText}</InputLabel>
+                <InputLabel id={labelId}>All Spells</InputLabel>
                 <Select
                     id={id}
-                    label={labelText}
+                    label="All Spells"
                     labelId={labelId}
                     onChange={handleRotateChange}
                     value={value}
                 >
                     {
-                        nearbyRotates[0]
-                        ? nearbyRotates.map(rotate => {
+                        allRotates[0]
+                        ? allRotates.map(rotate => {
                             return (
                                 <MenuItem
                                     key={rotate.id}
